Удаление ненужного модуля
